:bug: Cleans up Pytorch LAC code and fixes Pytorch Sac return maximization bug
I cleaned up the Pytorch version of the LAC algorithm . While doing this
I noticed that the SAC algorithm which can be enabled with "use_lyapunov=False" was still
maximizing instead of minimizing. For the environments we currently use
we need to minimize the error. This was already changed in all the other
versions.